+++ actionpack/CHANGELOG.md
@@ -1,4 +1,6 @@
-*   Add `action_dispatch.verbose_redirect_logs` setting that logs where redirects were called from.
+*   Deprecate `ActionController::MissingRenderer`
+
+    *zzak*
 
     Similar to `active_record.verbose_query_logs` and `active_job.verbose_enqueue_logs`, this adds a line in your logs that shows where a redirect was called from.